+++ 04_create_reference_channel_gold_table.py
@@ -2,20 +2,19 @@
 # MAGIC %md
 # MAGIC ### Overview - Create Reference Channel and Create Gold Table 
 # MAGIC
-# MAGIC While EEG recordings can be made without a reference electrode, the use of a reference electrode is essential for accurate and meaningful EEG analysis. It helps in grounding the electrical potentials, canceling out common-mode noise, facilitating signal comparison, and enabling various analytical techniques. The choice of reference scheme should be made based on the experimental requirements and analytical considerations.
+# MAGIC Applying a second-order Butterworth Filter to EEG signals allows you to isolate specific frequency components while reducing the strength or amplitude of others, thereby enabling certain frequency components to pass through relatively unaffected. This process aids in extracting meaningful information from EEG data, thereby enhancing the analysis of brain activity across various frequency bands.
+# MAGIC
+# MAGIC The EEG signals from each channel underwent filtering using a second-order Butterworth Filter across distinct physiological frequency ranges: 2–4 Hz (delta), 4.5–7.5 Hz (theta), 8–12.5 Hz (alpha), 13–30 Hz (beta), and 30–45 Hz (gamma).
 # MAGIC
 # MAGIC ##### In this notebook we will:
-# MAGIC   * Contrast the different techniques used in EEG data analysis to re-reference the data.
-# MAGIC     * Zero Reference Method
-# MAGIC       - In this method, you choose one electrode to be the reference and subtract its signal from all other electrodes.
-# MAGIC     * Average Reference Method
-# MAGIC       - In this method, you calculate the average signal across all electrodes and subtract this average from each electrode. This method assumes that the average potential of all electrodes represents a good approximation of zero potential.
+# MAGIC   * Isolate specific EEG frequencies using a second-order Butterworth Filter.
 # MAGIC   * Explore dataset based on summarize statistics. 
 # MAGIC   * Examine statistical metrics utilizing Databricks' integrated commands `describe` and `summary` for potential future data manipulation.
 # MAGIC     * `describe`: Provides statistics including count, mean, standard deviation, minimum, and maximum.
 # MAGIC     * `summary`: describe + interquartile range (IQR)
 # MAGIC     * Look for missing or null values in the data that will cause outliers or skew the data, zero values are expected with EEG data so no need to wrangle them.
 # MAGIC   * Create Gold Layer Table
+# MAGIC
 
 # COMMAND ----------
 
@@ -24,10 +23,16 @@
 
 # COMMAND ----------
 
-df_silver_all = spark.sql("""SELECT * FROM main.solution_accelerator.eeg_butter_filtered_silver""")
+df_zero_ref_all = spark.sql("""SELECT * FROM main.solution_accelerator.eeg_zero_ref_silver""")
+df_avg_ref_all = spark.sql("""SELECT * FROM main.solution_accelerator.eeg_avg_ref_silver""")
 
 # Show the result
-df_silver_all.show()
+display(df_zero_ref_all)
+# df_zero_ref_all.show()
+
+# Show the result
+display(df_avg_ref_all)
+# df_avg_ref_all.show()
 
 # COMMAND ----------
 
@@ -37,215 +42,225 @@ df_silver_all.show()
 # COMMAND ----------
 
 # Convert our two PySpark Dataframes to Pandas Dataframes
-display(df_silver_all.head())
+display(df_zero_ref_all.head())
 
-df_patients_all = df_silver_all.toPandas()
+df_zero_ref = df_zero_ref_all.toPandas()
 
-# COMMAND ----------
+# Convert our two PySpark Dataframes to Pandas Dataframes
+display(df_avg_ref_all.head())
 
-# MAGIC %md
-# MAGIC ##### Contrast the different techniques used in EEG data analysis to re-reference the data.
-# MAGIC
-# MAGIC Both the Zero Reference (ZR) and Average Reference (AR) methods are used in EEG (Electroencephalography) to mitigate common noise sources and spatial biases. However, they differ in their approach to reference point selection and signal processing. 
-# MAGIC
-# MAGIC Contrast the EEG reference points generated by employing both the ZR and AR methods to determine the perferred EEG reference point for our specific use case.
+df_avg_ref = df_avg_ref_all.toPandas()
 
 # COMMAND ----------
 
 # MAGIC %md
-# MAGIC ###### Zero Reference Method
-# MAGIC In this method, you choose one electrode to be the reference and subtract its signal from all other electrodes.
+# MAGIC ##### Isolate specific EEG frequencies using a second-order Butterworth Filter
 
 # COMMAND ----------
 
-# Helper library with many built-in functions
-%pip install mne
-
-# COMMAND ----------
-
-import os
-import mne
 import pandas as pd
-# import numpy as np
-from pyspark.sql.functions import lit
+import numpy as np
+from scipy import signal
+from scipy.signal import butter, filtfilt, freqz
+import matplotlib.pyplot as plt
 
-####### CALCULATING THE REFERENCE ELECTRODE USING THE ZERO REFERENCE METHOD #######
+def butter_bandpass(lowcut, highcut, fs, order=2):
+    nyq = 0.5 * fs
+    low = lowcut / nyq
+    high = highcut / nyq
+    b, a = butter(order, [low, high], btype='band')
+    return b, a
+
+def butter_bandpass_filter(data, lowcut, highcut, fs, order=2):
+    b, a = butter_bandpass(lowcut, highcut, fs, order=order)
+    y = filtfilt(b, a, data)
+    return y
+
+# Define frequency bands
+frequency_bands = {
+    'delta': (2, 4),
+    'theta': (4.5, 7.5),
+    'alpha': (8, 12.5),
+    'beta': (13, 30),
+    'gamma': (30, 45)
+}
+
+# EEG data parameters
+sampling_freq = 250  # frequency in Hz
+duration = 900  # Duration of EEG data in seconds
+num_samples = sampling_freq * duration
+num_channels = 19  # Number of EEG channels
+
+# Define filter parameters
+order = 2  # Filter order
+
+####### Butterworth Filter #######
+
+# For each patient, signals of each EEG channel were filtered
 
 # Make a copy of the Dataframe because we are going to apply two different methods and we dont want to override the original
-df_patients_all_cp = df_patients_all.copy()
+df_zero_ref_cp = df_zero_ref.copy()
 
-# TODO: Perform the following steps for each patient
+# Extract patient_id column
+pt_names = list(df_zero_ref_cp['patient_id'].unique())
+print(f"patient_names:::{pt_names}")
 
 filtered_data = {}
-for col in df_patients_all_cp.columns:
-    if col != 'patient_id':
-        filtered_data[col] = df_patients_all_cp.values
-    else:
-        # Keep the excluded column as is
-        filtered_data[col] = df_patients_all_cp[col].values 
 
-# Get channel values only, extract patient_id column. Numpy will remove the index and column labels
-df_numeric = df_patients_all_cp.select_dtypes(include=[np.number])
+for pt in pt_names:
+    print("PATIENT_NAME::", pt);
+    filtered_data[pt] = {}
+    for c_name in df_zero_ref_cp.head():
+        print("COLUMN_NAME::", c_name);
+        if c_name != 'patient_id':
+            filtered_data[pt][c_name] ={}
+            ch_data = df_zero_ref_cp[c_name].loc[df_zero_ref_cp['patient_id'] == pt]
 
-# Convert DataFrame to NumPy array and extract the values and transpose it (channels x samples)
-np_samples_array = df_numeric.values.T
+            for band, (lowcut, highcut) in frequency_bands.items():
+                # Get Patients Electrode Readings
 
-# Sampling rate in Hz
-sfreq = 250
+                print("---BAND::", band)
+                filtered_data[pt][c_name][band] = butter_bandpass_filter(ch_data.values, lowcut, highcut, sampling_freq, order)
+        # else:
+        #     # Keep the excluded column as is
+        #     filtered_data[c_name] = df_patients[c_name].values 
 
-# Get channel names , extract patient_id column
-ch_names = [c for c in df_silver_all.columns if c != 'patient_id']
-print(f"ref_channels:::{ch_names}")
+        # # Print filtered EEG data shape
+        # for band, data in filtered_data.items():
+        #     print(f'{band} band filtered data shape: {data.shape}')
 
-# Create an info structure needed by MNE
-info = mne.create_info(ch_names=ch_names, sfreq=sfreq, ch_types='eeg')
+        # # Compute the frequency response of the filter
+        # b, a = butter_bandpass(df_patients[c_name].values, lowcut, highcut, sampling_freq, order)
+        # w, h = freqz(b, a, worN=8000)
 
-# Create the MNE Raw object
-mne_raw = mne.io.RawArray(filtered_data[col], info)
+        # # Convert the normalized frequencies to Hz
+        # frequencies = 0.5 * sampling_freq * w / np.pi
 
-# Now we have our MNE Raw object and are ready for further analysis
+        # # Plot the frequency response
+        # plt.figure()
+        # plt.plot(frequencies, np.abs(h), 'b')
+        # plt.title('Butter Filter frequency response for EEG Data each subject.')
+        # plt.xlabel('Frequency (Hz)')
+        # plt.ylabel('Gain')
+        # plt.grid()
+        # plt.show()
 
-# Specify the reference channels, extract columns 'Fz' and 'Cz'
-ref_channels = [c for c in df_patients_all_cp.columns if c in ['Fz', 'Cz']]
+        # plt.semilogx(w, 20 * np.log10(abs(h)))
+        # plt.title('Butter Filter frequency response for EEG Data each subject.')
+        # plt.xlabel('Frequency [radians / second]')
+        # plt.ylabel('Amplitude [dB]')
+        # plt.margins(0, 0.1)
+        # plt.grid(which='both', axis='both')
+        # plt.axvline(100, color='green') # cutoff frequency
+        # plt.show()
 
-# Compute the average reference signal from Fz and Cz and apply it
-mne_raw.set_eeg_reference(ref_channels=ref_channels)
+        # t = np.linspace(0, 1, 1000, False)  # 1 second
+        # sig = np.sin(2*np.pi*10*t) + np.sin(2*np.pi*20*t)
+        # fig, (ax1, ax2) = plt.subplots(2, 1, sharex=True)
+        # ax1.plot(t, sig)
+        # ax1.set_title('10 Hz and 20 Hz sinusoids')
+        # ax1.axis([0, 1, -2, 2])
 
-# Extract EEG data from mne_raw and make a times column
-eeg_data, data_time = mne_raw.get_data(return_times=True)
-print(eeg_data.shape)
-print(data_time.shape)
+        # sos = signal.butter(10, 15, 'hp', fs=1000, output='sos')
+        # filtered = signal.sosfilt(sos, sig)
+        # ax2.plot(t, filtered)
+        # ax2.set_title('After 15 Hz high-pass filter')
+        # ax2.axis([0, 1, -2, 2])
+        # ax2.set_xlabel('Time [seconds]')
+        # plt.tight_layout()
+        # plt.title('Butter Filter frequency response for EEG Data each subject.')
+        # plt.show()
 
-# Create DataFrame
-df_zero_ref_all = pd.DataFrame(data=eeg_data.T, columns=ch_names)
 
-# Add a new column
-df_zero_ref_all['data_time'] = data_time
 
+# COMMAND ----------
+
+print(filtered_data.keys())
+for pt in filtered_data:
+    print("PT::", pt)
+    # print("KEYS:",filtered_data[pt].keys())
+    for ch in filtered_data[pt]:
+        print("---CH::", ch)
+        print(" ++ BANDS::", filtered_data[pt][ch].keys())
+        for band, data in filtered_data[pt][ch].items():
+            print(" +++ BAND::", band)
+            print(".  @@@@ DATA::", data);
+            print("     DATA LENGTH::", len(data))
+# TODO make an array of dataframes from the filtered data dictionary
+
+# Convert the filtered data dictionary into a DataFrame
+# df_butter_filtered = pd.DataFrame(filtered_data)
+
+# # Display the first few rows of the filtered DataFrame
+# print(df_butter_filtered.head())
+
+# COMMAND ----------
+
+# MAGIC %md
+# MAGIC ##### Delete if we run out of time - Finite Impulse Response (FIR) Filters Example
+
+# COMMAND ----------
+
+from scipy.signal import firwin, filtfilt
+import numpy as np
+import pandas as pd
+
+# TODO: Emergency, it makes sense to me that different patients would have different frequencies that need to be filtered out and this filter should be INDIVIDUAL dataset
+
+# Define the FIR bandpass filter functions
+def fir_bandpass(lowcut, highcut, fs, numtaps):
+    nyq = 0.5 * fs
+    taps = firwin(numtaps, [lowcut / nyq, highcut / nyq], pass_zero=False)
+    return taps
+
+def fir_bandpass_filter(data, lowcut, highcut, fs, numtaps):
+    taps = fir_bandpass(lowcut, highcut, fs, numtaps)
+    y = filtfilt(taps, [1.0], data)
+    return y
+
+# Remove columns that are not electrodes
+df_channels = df_patients.drop(columns=['patient_id'])
+
+# Filter parameters
+lowcut = 1.0   # Lower cutoff frequency in Hz
+highcut = 50.0 # Upper cutoff frequency in Hz
+fs = 250.0     # Sampling frequency in Hz
+numtaps = 101  # Number of filter taps (higher means sharper frequency response)
+
+# Apply the filter to each column in the DataFrame
+filtered_data = {}
+for column in df_channels.columns:
+    filtered_data[column] = fir_bandpass_filter(df_channels[column].values, lowcut, highcut, fs, numtaps)
+
+# Convert the filtered data dictionary to a DataFrame
+df_fir_filtered = pd.DataFrame(filtered_data)
+
+# Display the first few rows of the filtered DataFrame
+print(df_fir_filtered.head())
+
+
+# COMMAND ----------
+
+# MAGIC %md
+# MAGIC ###### Checking number of rows in DataFrame equals number of rows in original data
 
 # COMMAND ----------
 
 # Output the Dataframe Schema 
-df_zero_ref_all.printSchema()
+df_butter_filtered.printSchema()
 
 # Checking number of rows in Dataframes
-rows_all = df_zero_ref_all.count()
+rows_all = df_butter_filtered.count()
 print(f"Number of N rows {rows_all}")
 
 # COMMAND ----------
 
-# MAGIC %md
-# MAGIC ###### Average Reference Method
-# MAGIC This method re-references each electrode to the average potential of all electrodes.
-
-# COMMAND ----------
-
-# Install required libraries
-%pip install pyEDFlib
-
-# COMMAND ----------
-
-import os
-import mne
-import pandas as pd
-import pyedflib
-import numpy as np
-from pyspark.sql.functions import lit
-
-# Directory path
-directory = '/tmp'
-
-# Iterate over files in the directory
-for filename in os.listdir(directory):
-    print(filename)
-    # Path to the EEG file (EDF format file in DBFS)
-    if os.path.isfile(os.path.join(directory, filename)):
-        if "edf" in filename:
-            # Create the file path
-            filepath = os.path.join(directory, filename)
-
-            # Read the EDF file
-            edf_f = pyedflib.EdfReader(filepath)
-
-            # Extract data from the .edf file
-
-            # Get the number of signals, i.e. number of electrode locations
-            n_signals = edf_f.signals_in_file
-
-            # Get the labels of each signal, i.e. electrode location names
-            signal_labels = edf_f.getSignalLabels()
-            
-            print(f"signal_labels:::{signal_labels}")
-
-            # Create a dictionary to hold the data
-            signal_data = {}
-
-            # Read each signal into the dictionary
-            for i in range(n_signals):
-                signal_data[signal_labels[i]] = edf_f.readSignal(i)
-                # Verify we have 250 Hz of datapoints, 15 min * 60 sec = 900, 900 * 250 = 225000
-                print(f"Num of Signals:::{signal_labels[i]} Num{len( signal_data[signal_labels[i]])}")
-            print(signal_data.keys())
-
-            # Close the .edf file
-            edf_f._close()
-            del edf_f
-
-            # making a Pandas DataFrame first is faster and will ensure the double data type for the signals 
-            df_eeg_data = pd.DataFrame.from_dict(signal_data)
-
-            ####### CALCULATING THE REFERENCE ELECTRODE USING THE AVERAGE REFERENCE METHOD #######
-
-            # Calculate the average signal across all channels
-            average_reference = np.mean(df_eeg_data, axis=0)
-
-            # Subtract the average reference from each channel
-            eeg_data_average_referenced = df_eeg_data - average_reference
-
-            # eeg_data_average_referenced is now re-referenced to the average reference
-
-            fn = filename.replace(".edf","")
-            print(f"FN::{fn}")
-
-            df_spark = spark.createDataFrame(eeg_data_average_referenced)
-            df_spark = df_spark.withColumn('patient_id', lit(fn))
-     
-            pt = "h"
-            if fn.startswith("s"):
-                pt = "s"
-            df_avg_ref_subjects_all = df_spark.withColumn('subject', lit(pt))
-
-# COMMAND ----------
-
-# MAGIC %md
-# MAGIC ##### Average Reference Method
-
-# COMMAND ----------
-
-import mne
-
-####### CALCULATING THE REFERENCE ELECTRODE USING THE AVERAGE REFERENCE METHOD #######
-
-# Make a copy of the Dataframe because we are going to apply two different methods and we dont want to override the original
-df_patients_all_cp = df_patients_all.copy()
-
-raw = mne.io.read_raw_fif(df_patients_all_cp, preload=True)
-
-# Apply average reference
-df_avg_ref_all = raw.copy().set_eeg_reference(ref_channels='average')
-
-# Plot the average referenced data
-df_avg_ref_all.plot(n_channels=10, title='Average Referenced Data', show=True)
-
-
-# COMMAND ----------
-
 # Output the Dataframe Schema 
-df_avg_ref_all.printSchema()
+df_fir_filtered.printSchema()
 
 # Checking number of rows in Dataframes
-rows_sub_all = df_avg_ref_all.count()
-print(f"Number of N rows {rows_sub_all}")
+rows_all = df_fir_filtered.count()
+print(f"Number of N rows {rows_all}")
 
 # COMMAND ----------
 
@@ -254,8 +269,8 @@ print(f"Number of N rows {rows_sub_all}")
 
 # COMMAND ----------
 
-dbutils.data.summarize(df_zero_ref_all)
-dbutils.data.summarize(df_avg_ref_all)
+dbutils.data.summarize(df_butter_filtered)
+dbutils.data.summarize(df_fir_filtered)
 
 # COMMAND ----------
 
@@ -264,26 +279,26 @@ dbutils.data.summarize(df_avg_ref_all)
 
 # COMMAND ----------
 
-df_zero_ref_all.describe()
-df_avg_ref_all.describe()
+df_butter_filtered.describe()
+df_fir_filtered.describe()
 
 # COMMAND ----------
 
-df_zero_ref_all.summary()
-df_avg_ref_all.summary()
+df_butter_filtered.summary()
+df_fir_filtered.summary()
 
 # COMMAND ----------
 
 # MAGIC %md
 # MAGIC ##### Create Gold Layer Table
+# MAGIC ###### Use Butterworth Filter data
 
 # COMMAND ----------
 
 # MAGIC %sql
 # MAGIC -- Dropping the table because we may have updated the Dataframe
 # MAGIC
-# MAGIC DROP TABLE IF EXISTS main.solution_accelerator.eeg_zero_ref_gold;
-# MAGIC DROP TABLE IF EXISTS main.solution_accelerator.eeg_avg_ref_gold;
+# MAGIC DROP TABLE IF EXISTS main.solution_accelerator.eeg_butter_filtered_gold;
 
 # COMMAND ----------
 
@@ -292,11 +307,14 @@ df_avg_ref_all.summary()
 # Establish a persistent delta table by converting the previously created Spark DataFrames into a Delta Tables.
 
 # Replace any previously existing table and register the DataFrame as a Delta table in the metastore.
-df_zero_ref_all.write.format("delta").mode("overwrite").saveAsTable("main.solution_accelerator.eeg_zero_ref_data_gold")
-df_avg_ref_all.write.format("delta").mode("overwrite").saveAsTable("main.solution_accelerator.eeg_avg_ref_data_gold")
+df_butter_filtered.write.format("delta").mode("overwrite").saveAsTable("main.solution_accelerator.eeg_butter_filtered_gold")
 
 print("Tables exist.")
 
 # Delete the Dataframes from memory
-del df_zero_ref_all
-del df_avg_ref_all
+del df_butter_filtered
+
+# COMMAND ----------
+
+# MAGIC %md
+# MAGIC
